# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Kiali project, observability for the Istio service mesh",
    "title": "Kiali",
    "version": "_"
  },
  "basePath": "/api",
  "paths": {
...
    "/namespaces/{namespace}/services": {
      "get": {
        "description": "Endpoint to get the details of a given service",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "services"
        ],
        "operationId": "serviceList",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The namespace name.",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/serviceListResponse"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
    "/namespaces/{namespace}/services/{service}": {
      "get": {
        "description": "Endpoint to get the details of a given service",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "services"
        ],
        "operationId": "serviceDetails",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The namespace name.",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The service name.",
            "name": "service",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/serviceDetailsResponse"
          },
          "404": {
            "$ref": "#/responses/notFoundError"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
    "/namespaces/{namespace}/services/{service}/dashboard": {
      "get": {
        "description": "Endpoint to fetch dashboard to be displayed, related to a single service",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "services"
        ],
        "operationId": "serviceDashboard",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The namespace name.",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The service name.",
            "name": "service",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "default": "true",
            "x-go-name": "Name",
            "description": "Flag for fetching histogram average. Default is true.",
            "name": "avg",
            "in": "query"
          },
          {
            "type": "string",
            "default": "[]",
            "x-go-name": "Name",
            "description": "List of labels to use for grouping metrics (via Prometheus 'by' clause).",
            "name": "byLabels[]",
            "in": "query"
          },
          {
            "type": "string",
            "default": "outbound",
            "x-go-name": "Name",
            "description": "Traffic direction: 'inbound' or 'outbound'.",
            "name": "direction",
            "in": "query"
          },
          {
            "type": "string",
            "default": "1800",
            "x-go-name": "Name",
            "description": "Duration of the query period, in seconds.",
            "name": "duration",
            "in": "query"
          },
          {
            "type": "string",
            "default": "[]",
            "x-go-name": "Name",
            "description": "List of quantiles to fetch. Fetch no quantiles when empty. Ex: [0.5, 0.95, 0.99].",
            "name": "quantiles[]",
            "in": "query"
          },
          {
            "type": "string",
            "default": "rate",
            "x-go-name": "Name",
            "description": "Prometheus function used to calculate rate: 'rate' or 'irate'.",
            "name": "rateFunc",
            "in": "query"
          },
          {
            "type": "string",
            "default": "1m",
            "x-go-name": "Name",
            "description": "Interval used for rate and histogram calculation.",
            "name": "rateInterval",
            "in": "query"
          },
          {
            "type": "string",
            "default": "all protocols",
            "x-go-name": "Name",
            "description": "Desired request protocol for the telemetry: For example, 'http' or 'grpc'.",
            "name": "requestProtocol",
            "in": "query"
          },
          {
            "type": "string",
            "default": "source",
            "x-go-name": "Name",
            "description": "Istio telemetry reporter: 'source' or 'destination'.",
            "name": "reporter",
            "in": "query"
          },
          {
            "type": "string",
            "default": "15",
            "x-go-name": "Name",
            "description": "Step between [graph] datapoints, in seconds.",
            "name": "step",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/dashboardResponse"
          },
          "400": {
            "$ref": "#/responses/badRequestError"
          },
          "503": {
            "$ref": "#/responses/serviceUnavailableError"
          }
        }
      }
    },
...
    "/threescale": {
      "get": {
        "description": "Endpoint to check if threescale adapter is present in the cluster and if user can write adapter config",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "operationId": "getThreeScaleInfo",
        "responses": {
          "200": {
            "$ref": "#/responses/threeScaleInfoResponse"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
    "/threescale/handlers": {
      "get": {
        "description": "Endpoint to fetch threescale handlers generated from Kiali",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "operationId": "getThreeScaleHandlers",
        "responses": {
          "200": {
            "$ref": "#/responses/threeScaleHandlersResponse"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      },
      "post": {
        "description": "Endpoint to create a new threescale handler+instance generated by Kiali",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "operationId": "postThreeScaleHandlers",
        "responses": {
          "200": {
            "$ref": "#/responses/threeScaleHandlersResponse"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
    "/threescale/handlers/{threescaleHandlerName}": {
      "delete": {
        "description": "Endpoint to delete an existing threescale handler+instance generated by Kiali",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "operationId": "deleteThreeScaleHandler",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The ThreeScaleHandler name.",
            "name": "threescaleHandlerName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/threeScaleHandlersResponse"
          },
          "404": {
            "$ref": "#/responses/notFoundError"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      },
      "patch": {
        "description": "Endpoint to update an existing threescale handler generated by Kiali",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "operationId": "patchThreeScaleHandler",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Name",
            "description": "The ThreeScaleHandler name.",
            "name": "threescaleHandlerName",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/threeScaleHandlersResponse"
          },
          "404": {
            "$ref": "#/responses/notFoundError"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Addresses": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Address"
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
    "Aggregation": {
      "description": "Aggregation is the model representing label's allowed aggregation, transformed from aggregation in MonitoringDashboard k8s resource",
      "type": "object",
      "properties": {
        "displayName": {
          "type": "string",
          "x-go-name": "DisplayName"
        },
        "label": {
          "type": "string",
          "x-go-name": "Label"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "Chart": {
      "description": "Chart is the model representing a custom chart, transformed from charts in MonitoringDashboard k8s resource",
      "type": "object",
      "properties": {
        "histogram": {
          "$ref": "#/definitions/Histogram"
        },
        "metric": {
          "$ref": "#/definitions/Metric"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "spans": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Spans"
        },
        "unit": {
          "type": "string",
          "x-go-name": "Unit"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "Endpoint": {
      "type": "object",
      "properties": {
        "addresses": {
          "$ref": "#/definitions/Addresses"
        },
        "ports": {
          "$ref": "#/definitions/Ports"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
    "Endpoints": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Endpoint"
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "IstioValidations": {
      "title": "IstioValidations represents a set of IstioValidation grouped by IstioValidationKey.",
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "MTLSStatus": {
      "description": "MTLSStatus describes the current mTLS status of a mesh entity",
      "type": "object",
      "required": [
        "status"
      ],
      "properties": {
        "status": {
          "description": "mTLS status: MTLS_ENABLED, MTLS_PARTIALLY_ENABLED, MTLS_NOT_ENABLED",
          "type": "string",
          "x-go-name": "Status",
          "example": "MTLS_ENABLED"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "MonitoringDashboard": {
      "description": "MonitoringDashboard is the model representing custom monitoring dashboard, transformed from MonitoringDashboard k8s resource",
      "type": "object",
      "properties": {
        "aggregations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Aggregation"
          },
          "x-go-name": "Aggregations"
        },
        "charts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Chart"
          },
          "x-go-name": "Charts"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "Ports": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Port"
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "ResourcePermissions": {
      "description": "ResourcePermissions holds permission flags for an object type\nTrue means allowed.",
      "type": "object",
      "properties": {
        "create": {
          "type": "boolean",
          "x-go-name": "Create"
        },
        "delete": {
          "type": "boolean",
          "x-go-name": "Delete"
        },
        "update": {
          "type": "boolean",
          "x-go-name": "Update"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "Service": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "externalName": {
          "type": "string",
          "x-go-name": "ExternalName"
        },
        "ip": {
          "type": "string",
          "x-go-name": "Ip"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Labels"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "$ref": "#/definitions/namespace"
        },
        "ports": {
          "$ref": "#/definitions/Ports"
        },
        "resourceVersion": {
          "type": "string",
          "x-go-name": "ResourceVersion"
        },
        "selectors": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Selectors"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
    "ServiceDetails": {
      "type": "object",
      "properties": {
        "destinationRules": {
          "$ref": "#/definitions/destinationRules"
        },
        "endpoints": {
          "$ref": "#/definitions/Endpoints"
        },
        "errorTraces": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorTraces"
        },
        "health": {
          "$ref": "#/definitions/ServiceHealth"
        },
        "istioSidecar": {
          "type": "boolean",
          "x-go-name": "IstioSidecar"
        },
        "namespaceMTLS": {
          "$ref": "#/definitions/MTLSStatus"
        },
        "service": {
          "$ref": "#/definitions/Service"
        },
        "validations": {
          "$ref": "#/definitions/IstioValidations"
        },
        "virtualServices": {
          "$ref": "#/definitions/virtualServices"
        },
        "workloads": {
          "$ref": "#/definitions/WorkloadOverviews"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "ServiceHealth": {
      "description": "ServiceHealth contains aggregated health from various sources, for a given service",
      "type": "object",
      "properties": {
        "requests": {
          "$ref": "#/definitions/RequestHealth"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
    "ServiceList": {
      "type": "object",
      "properties": {
        "namespace": {
          "$ref": "#/definitions/namespace"
        },
        "services": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServiceOverview"
          },
          "x-go-name": "Services"
        },
        "validations": {
          "$ref": "#/definitions/IstioValidations"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
    "ServiceOverview": {
      "type": "object",
      "required": [
        "name",
        "istioSidecar",
        "appLabel"
      ],
      "properties": {
        "appLabel": {
          "description": "Has label app",
          "type": "boolean",
          "x-go-name": "AppLabel",
          "example": true
        },
        "istioSidecar": {
          "description": "Define if Pods related to this Service has an IstioSidecar deployed",
          "type": "boolean",
          "x-go-name": "IstioSidecar",
          "example": true
        },
        "name": {
          "description": "Name of the Service",
          "type": "string",
          "x-go-name": "Name",
          "example": "reviews-v1"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "ThreeScaleHandler": {
      "description": "ThreeScaleHAndler represents the minimal info that a user needs to know from the UI to link a service with 3Scale site",
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "x-go-name": "AccessToken"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "serviceId": {
          "type": "string",
          "x-go-name": "ServiceId"
        },
        "systemUrl": {
          "type": "string",
          "x-go-name": "SystemUrl"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
    "ThreeScaleHandlers": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/ThreeScaleHandler"
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
    "ThreeScaleInfo": {
      "description": "ThreeScaleInfo shows if 3scale adapter is enabled in cluster and if user has permissions on adapter's configuration",
      "type": "object",
      "properties": {
        "enabled": {
          "type": "boolean",
          "x-go-name": "Enabled"
        },
        "permissions": {
          "$ref": "#/definitions/ResourcePermissions"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "WorkloadOverviews": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/WorkloadListItem"
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "destinationRules": {
      "description": "This is used for returning an array of DestinationRules",
      "type": "object",
      "title": "DestinationRules destinationRules",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/destinationRule"
          },
          "x-go-name": "Items"
        },
        "permissions": {
          "$ref": "#/definitions/ResourcePermissions"
        }
      },
      "x-go-name": "DestinationRules",
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "namespace": {
      "description": "A Namespace provide a scope for names\nThis type is used to describe a set of objects.",
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "description": "The id of the namespace.",
          "type": "string",
          "x-go-name": "Name",
          "example": "istio-system"
        }
      },
      "x-go-name": "Namespace",
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "virtualServices": {
      "description": "This type is used for returning an array of VirtualServices with some permission flags",
      "type": "object",
      "title": "VirtualServices virtualServices",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/virtualService"
          },
          "x-go-name": "Items"
        },
        "permissions": {
          "$ref": "#/definitions/ResourcePermissions"
        }
      },
      "x-go-name": "VirtualServices",
      "x-go-package": "github.com/mayadata-io/kiali/models"
    }
  },
  "responses": {
...
    "badRequestError": {
      "description": "BadRequestError: the client request is incorrect",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 400,
            "x-go-name": "Code",
            "example": 400
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "dashboardResponse": {
      "description": "Dashboard response model",
      "schema": {
        "$ref": "#/definitions/MonitoringDashboard"
      }
    },
...
    "internalError": {
      "description": "A Internal is the error message that means something has gone wrong",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 500,
            "x-go-name": "Code",
            "example": 500
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
    "notFoundError": {
      "description": "A NotFoundError is the error message that is generated when server could not find what was requested.",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 404,
            "x-go-name": "Code",
            "example": 404
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "serviceDetailsResponse": {
      "description": "Listing all the information related to a workload",
      "schema": {
        "$ref": "#/definitions/ServiceDetails"
      }
    },
...
    "serviceListResponse": {
      "description": "Listing all services in the namespace",
      "schema": {
        "$ref": "#/definitions/ServiceList"
      }
    },
    "serviceUnavailableError": {
      "description": "A Internal is the error message that means something has gone wrong",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 503,
            "x-go-name": "Code",
            "example": 503
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
    "threeScaleHandlersResponse": {
      "description": "List of ThreeScale handlers created from Kiali to be used in the adapter's configuration",
      "schema": {
        "$ref": "#/definitions/ThreeScaleHandlers"
      }
    },
    "threeScaleInfoResponse": {
      "description": "Return if ThreeScale adapter is enabled in Istio and if user has permissions to write adapter's configuration",
      "schema": {
        "$ref": "#/definitions/ThreeScaleInfo"
      }
    },
...
  }
}